Minor fixes.
Variable setting fixes so they consider passed conditionals.

Changed height calculation. Not sure how we want to handle this for content. In most use cases the video is used for a hero, which works great and calculates height.

I don’t want the width and height passed. Width is fine, but height should be calculated off of width, not a user input, in order to retain proper ratio.

Applied default 16:9. Nothing else is currently be accounted for- having 4:3 the default was silly, should use the most common format.

TODO: Add other aspect ratios or detect what is passed. Detection does work, but haven’t been able to properly test.